refactor(room-escape): 패키지 이름 변경
