+++ README.md
@@ -58,8 +58,16 @@
     - [X] 예약 조회) ID가 잘못된 경우 (float, string)
     - [X] 예약 삭제) ID가 없는 경우 삭제 불가
     - [X] 예약 삭제) ID가 잘못된 경우 (float, string)
+- [ ] 콘솔 애플리케이션에 데이터베이스를 적용한다.
+    - [ ] 직접 커넥션을 만들어서 데이터베이스에 접근한다.
+- [ ] 웹 애플리케이션에 데이터베이스를 적용한다.
+    -  [ ] 스프링이 제공하는 기능을 활용하여 데이터베이스에 접근한다.
 
 ## 프로그래밍 요구사항
 
 - 기존 콘솔 애플리케이션은 그대로 잘 동작해야한다.
 - 콘솔 애플리케이션과 웹 애플리케이션의 중복 코드는 허용한다. (다음 단계에서 리팩터링 예정)
+- 콘솔 애플리케이션과 웹 애플리케이션에서 각각 데이터베이스에 접근하는 로직을 구현한다.
+- 콘솔 애플리케이션에서 데이터베이스에 접근 시 JdbcTemplate를 사용하지 않는다. 직접 Connection을 생성하여 데이터베이스에 접근한다.
+- 웹 애플리케이션에서 데이터베이스 접근 시 JdbcTemplate를 사용한다.
+- 콘솔 애플리케이션과 웹 애플리케이션의 중복 코드 제거해본다.
