+++ src/main/java/roomescape/RoomEscapeApplication.java
@@ -1,12 +1,13 @@
-package nextstep;
+package roomescape;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RoomEscapeSpringApplication {
+public class RoomEscapeApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(RoomEscapeSpringApplication.class, args);
+        SpringApplication.run(RoomEscapeApplication.class, args);
     }
 }
+
