+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,11 +1,8 @@
-package nextstep.mapping;
+package roomescape.controller;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import nextstep.domain.Reservation;
-import nextstep.domain.Theme;
-import nextstep.exception.BadRequestException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,6 +12,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.Reservation;
+import roomescape.dto.Theme;
+import roomescape.exception.BadRequestException;
 
 
 @RestController
