+++ src/main/java/roomescape/dto/Reservation.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
