+++ src/main/java/roomescape/dto/Theme.java
@@ -1,6 +1,7 @@
-package nextstep.domain;
+package roomescape.dto;
 
 public class Theme {
+
     private String name;
     private String desc;
     private Integer price;
