+++ src/main/java/roomescape/exception/BadRequestException.java
@@ -1,4 +1,4 @@
-package nextstep.exception;
+package roomescape.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
