+++ src/test/java/roomescape/controller/ReservationTest.java
@@ -1,11 +1,10 @@
-package nextstep.mapping;
+package roomescape.controller;
 
 import static org.hamcrest.core.Is.is;
 
 import io.restassured.RestAssured;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import nextstep.domain.Reservation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,6 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import roomescape.dto.Reservation;
 
 
 @DisplayName("웹 요청 / 응답 처리로 입출력 추가")
